backend/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
httpx = "^0.28.1"

[project]
name = "3-schach"
version = "0.1.0"
description = ""
authors = [
    {name = "lee"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "chess (>=1.11.2,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pymongo (>=4.8.0,<5.0.0)",
    "motor (>=3.5.0,<4.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "aio-pika (>=9.4.0,<10.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.23.0,<1.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
